requirements.txt
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.3
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
Flask-Testing==0.8.1
flask-wtf==1.2.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
infinity==1.5
iniconfig==2.0.0
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
tomli==2.0.1
typing-extensions==4.8.0
urllib3==2.1.0
validators==0.22.0
werkzeug==3.0.1
wtforms==3.1.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
zipp==3.17.0
